ran code cleanup and renamed like 5 things,  200+ files changed

+++ Biomes/CrabCrevice/CrabCreviceDroplet.cs
@@ -39,7 +39,7 @@ namespace Aequus.Biomes.CrabCrevice
                 {
                     frameDuration = 24 + Main.rand.Next(96);
                 }
-                if ((int)gore.frameCounter >= frameDuration)
+                if (gore.frameCounter >= frameDuration)
                 {
                     gore.frameCounter = 0;
                     gore.frame += 1;
@@ -54,7 +54,7 @@ namespace Aequus.Biomes.CrabCrevice
             else if (gore.frame <= 6)
             {
                 frameDuration = 8;
-                if ((int)gore.frameCounter >= frameDuration)
+                if (gore.frameCounter >= frameDuration)
                 {
                     gore.frameCounter = 0;
                     gore.frame += 1;
@@ -76,7 +76,7 @@ namespace Aequus.Biomes.CrabCrevice
                 {
                     gore.velocity.Y = 12f;
                 }
-                if ((int)gore.frameCounter >= frameDuration)
+                if (gore.frameCounter >= frameDuration)
                 {
                     gore.frameCounter = 0;
                     gore.frame += 1;
@@ -89,7 +89,7 @@ namespace Aequus.Biomes.CrabCrevice
             else
             {
                 gore.velocity.Y += 0.1f;
-                if ((int)gore.frameCounter >= frameDuration)
+                if (gore.frameCounter >= frameDuration)
                 {
                     gore.frameCounter = 0;
                     gore.frame += 1;
@@ -125,7 +125,7 @@ namespace Aequus.Biomes.CrabCrevice
                 if (Main.tile[tileX, tileY].LiquidAmount > 0)
                 {
                     gore.velocity *= 0f;
-                    gore.position.Y = tileY * 16 - (int)(Main.tile[tileX, tileY].LiquidAmount / 16);
+                    gore.position.Y = tileY * 16 - Main.tile[tileX, tileY].LiquidAmount / 16;
                 }
             }
